Change current hubs page to home page, and home page to hub page
