+++ next.config.js
@@ -20,25 +20,21 @@ const nextConfig = {
   },
   async rewrites() {
     return [
-      { source: '/my-chats', destination: '/hubs' },
-      { source: '/hot-chats', destination: '/hubs' },
+      { source: '/hubs', destination: '/' },
+      { source: '/my-chats', destination: '/' },
+      { source: '/hot-chats', destination: '/' },
     ]
   },
   async redirects() {
     return [
-      {
-        source: '/',
-        destination: '/x',
-        permanent: false,
-      },
       {
         source: '/c/:slug',
         destination: '/x/c/:slug',
         permanent: false,
       },
       {
-        source: '/:spaceId/c/:slug',
-        destination: '/:spaceId/:slug',
+        source: '/:hubId/c/:slug',
+        destination: '/:hubId/:slug',
         permanent: false,
       },
     ]
