+++ src/modules/chat/HomePage/HomePage.tsx
@@ -1,49 +1,123 @@
-import ChatPreviewList from '@/components/chats/ChatPreviewList'
-import NoChatsFound from '@/components/chats/NoChatsFound'
+import ChatSpecialButtons from '@/components/chats/ChatSpecialButtons'
 import DefaultLayout from '@/components/layouts/DefaultLayout'
+import NavbarWithSearch from '@/components/navbar/Navbar/custom/NavbarWithSearch'
+import Tabs, { TabsProps } from '@/components/Tabs'
 import useIsInIframe from '@/hooks/useIsInIframe'
 import useSearch from '@/hooks/useSearch'
+import { getMainHubId } from '@/utils/env/client'
+import { replaceUrl } from '@/utils/window'
 import dynamic from 'next/dynamic'
-import useSortedChats from '../hooks/useSortedChats'
-import HomePageNavbar from './HomePageNavbar'
+import { useEffect, useState } from 'react'
+import HotChatsContent from './HotChatsContent'
+import HubsContent from './HubsContent'
+import MyChatsContent from './MyChatsContent'
 
 const WelcomeModal = dynamic(() => import('@/components/modals/WelcomeModal'), {
   ssr: false,
 })
 
-export type HomePageProps = {
-  hubId: string
+export type HubsPageProps = {
+  isIntegrateChatButtonOnTop: boolean
+  hubsChatCount: { [id: string]: number }
 }
-const searchKeys = ['content.title']
-export default function HomePage({ hubId }: HomePageProps) {
+
+export type CommonHubContentProps = {
+  getSearchResults: ReturnType<typeof useSearch>['getSearchResults']
+}
+
+const hotChatsHubId = getMainHubId()
+export default function HubsPage(props: HubsPageProps) {
   const isInIframe = useIsInIframe()
+  const { search, setSearch, getSearchResults, focusController } = useSearch()
 
-  const { chats, allChatIds } = useSortedChats(hubId)
+  const renderHubsContent = (
+    children: JSX.Element,
+    showSpecialButtons?: boolean
+  ) => {
+    return (
+      <HubsContentWrapper
+        search={search}
+        isIntegrateChatButtonOnTop={props.isIntegrateChatButtonOnTop}
+        showSpecialButtons={showSpecialButtons}
+      >
+        {children}
+      </HubsContentWrapper>
+    )
+  }
 
-  const { search, getSearchResults, setSearch, focusController } = useSearch()
-  const { focusedElementIndex, searchResults } = getSearchResults(
-    chats,
-    searchKeys
-  )
+  const tabs: TabsProps['tabs'] = [
+    {
+      id: 'my-chats',
+      text: 'My Chats',
+      content: (setSelectedTab) =>
+        renderHubsContent(
+          <MyChatsContent
+            changeTab={setSelectedTab}
+            search={search}
+            getSearchResults={getSearchResults}
+          />
+        ),
+    },
+    {
+      id: 'hot-chats',
+      text: 'Hot Chats',
+      content: () =>
+        renderHubsContent(
+          <HotChatsContent
+            getSearchResults={getSearchResults}
+            hubId={hotChatsHubId}
+          />
+        ),
+    },
+    {
+      id: 'hubs',
+      text: 'Hubs',
+      content: () =>
+        renderHubsContent(
+          <HubsContent
+            getSearchResults={getSearchResults}
+            hubsChatCount={props.hubsChatCount}
+          />,
+          true
+        ),
+    },
+  ]
+
+  const [isTabUrlLoaded, setIsTabUrlLoaded] = useState(false)
+  const [selectedTab, setSelectedTab] = useState(1)
+  useEffect(() => {
+    const currentPathname = window.location.pathname.substring(1)
+
+    const index = tabs.findIndex(({ id }) => id === currentPathname)
+    if (index > -1) setSelectedTab(index)
+
+    setIsTabUrlLoaded(true)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
+  const usedSelectedTab = isTabUrlLoaded ? selectedTab : -1
+  const usedSetSelectedTab = (selectedTab: number) => {
+    setSelectedTab(selectedTab)
+    const selectedTabId = tabs[selectedTab]?.id
+    if (selectedTabId) replaceUrl(`/${selectedTabId}`)
+  }
 
   return (
     <DefaultLayout
       navbarProps={{
-        backButtonProps: { defaultBackLink: '/hubs' },
-        customContent: ({
-          backButton,
-          logoLink,
-          authComponent,
-          colorModeToggler,
-        }) => {
+        customContent: ({ logoLink, authComponent, colorModeToggler }) => {
           return (
-            <HomePageNavbar
-              chatsCount={allChatIds.length}
-              auth={authComponent}
-              colorModeToggler={colorModeToggler}
-              backButton={backButton}
-              logo={logoLink}
-              spaceId={hubId}
+            <NavbarWithSearch
+              customContent={(searchButton) => (
+                <div className='flex w-full items-center justify-between gap-4'>
+                  {logoLink}
+                  <div className='flex items-center gap-2'>
+                    {searchButton}
+                    {colorModeToggler}
+                    <div className='ml-1.5'>{authComponent}</div>
+                  </div>
+                </div>
+              )}
               searchProps={{
                 search,
                 setSearch,
@@ -54,16 +128,42 @@ export default function HomePage({ hubId }: HomePageProps) {
         },
       }}
     >
+      <Tabs
+        className='border-b border-border-gray bg-background-light md:bg-background-light/50'
+        panelClassName='mt-0 px-0'
+        asContainer
+        tabs={tabs}
+        withHashIntegration={false}
+        manualTabControl={{
+          selectedTab: usedSelectedTab,
+          setSelectedTab: usedSetSelectedTab,
+        }}
+      />
       {!isInIframe && <WelcomeModal />}
-      <div className='flex flex-col'>
-        {searchResults.length === 0 && (
-          <NoChatsFound search={search} hubId={hubId} />
-        )}
-        <ChatPreviewList
-          chats={searchResults}
-          focusedElementIndex={focusedElementIndex}
-        />
-      </div>
     </DefaultLayout>
+  )
+}
+
+function HubsContentWrapper({
+  isIntegrateChatButtonOnTop,
+  children,
+  search,
+  showSpecialButtons,
+}: {
+  isIntegrateChatButtonOnTop: boolean
+  children: JSX.Element
+  search: string
+  showSpecialButtons?: boolean
+}) {
+  const isInIframe = useIsInIframe()
+  return (
+    <div className='flex flex-col'>
+      {showSpecialButtons && !isInIframe && !search && (
+        <ChatSpecialButtons
+          isIntegrateChatButtonOnTop={isIntegrateChatButtonOnTop}
+        />
+      )}
+      {children}
+    </div>
   )
 }
