+++ src/modules/chat/HubPage/index.tsx
@@ -1,0 +1,4 @@
+import HubPage from './HubPage'
+
+export * from './HubPage'
+export default HubPage
