--- src/modules/chat/HubsPage/index.tsx
@@ -1,3 +1,0 @@
-import HubsPage from './HubsPage'
-
-export default HubsPage
