+++ src/pages/[hubId]/[slug]/about.tsx
