+++ src/pages/[hubId]/[slug]/index.tsx
@@ -1,5 +1,5 @@
 import { CHAT_PER_PAGE } from '@/constants/chat'
-import { getSpaceIdFromAlias } from '@/constants/chat-room'
+import { getHubIdFromAlias } from '@/constants/hubs'
 import ChatPage, { ChatPageProps } from '@/modules/chat/ChatPage'
 import { getPostsFromCache } from '@/pages/api/posts'
 import { AppCommonProps } from '@/pages/_app'
@@ -50,12 +50,12 @@ export const getStaticProps = getCommonStaticProps<
     head: { disableZoom: true },
   }),
   async (context) => {
-    const spaceIdOrAlias = context.params?.spaceId as string
+    const hubIdOrAlias = context.params?.hubId as string
     const slugParam = context.params?.slug as string
     const chatId = getValidatedChatId(slugParam)
     if (!chatId) return undefined
 
-    const spaceId = getSpaceIdFromAlias(spaceIdOrAlias) || spaceIdOrAlias
+    const hubId = getHubIdFromAlias(hubIdOrAlias) || hubIdOrAlias
 
     const queryClient = new QueryClient()
 
@@ -64,7 +64,7 @@ export const getStaticProps = getCommonStaticProps<
     try {
       const [{ messageIds, messages, chatData }] = await Promise.all([
         getChatsData(chatId),
-        prefetchBlockedEntities(queryClient, spaceId, [chatId]),
+        prefetchBlockedEntities(queryClient, hubId, [chatId]),
       ] as const)
 
       title = chatData?.content?.title || null
@@ -86,7 +86,7 @@ export const getStaticProps = getCommonStaticProps<
       props: {
         dehydratedState: dehydrate(queryClient),
         chatId,
-        hubId: spaceId,
+        hubId: hubId,
         head: {
           title,
           description: desc,
